Began development on documentation, fixed pow issue
